+++ nat-lab/tests/test_telio_features.py
@@ -37,10 +37,12 @@ def test_telio_features():
             qos=Qos(rtt_interval=5, rtt_tries=3, rtt_types=["Ping"], buckets=5),
             heartbeat_interval=3600,
             initial_heartbeat_interval=10,
+            enable_relay_conn_data=False,
+            enable_nat_type_collection=False,
         )
     )
     expected_nurse_qos = TelioFeatures.from_json(
-        """{"is_test_env": true, "exit_dns": {"auto_switch_dns_ips": true}, "nurse": {"fingerprint": "fingerprint", "qos": {"rtt_interval": 5, "rtt_tries": 3, "rtt_types": ["Ping"], "buckets": 5}, "heartbeat_interval": 3600, "initial_heartbeat_interval": 10}}"""
+        """{"is_test_env": true, "exit_dns": {"auto_switch_dns_ips": true}, "nurse": {"fingerprint": "fingerprint", "qos": {"rtt_interval": 5, "rtt_tries": 3, "rtt_types": ["Ping"], "buckets": 5}, "heartbeat_interval": 3600, "initial_heartbeat_interval": 10, "enable_nat_type_collection": false, "enable_relay_conn_data": false}}"""
     )
     assert nurse_qos_features == expected_nurse_qos
     assert nurse_qos_features.to_json() == expected_nurse_qos.to_json()
@@ -54,6 +56,8 @@ def test_telio_features():
             qos=Qos(rtt_interval=5, rtt_tries=3, rtt_types=["Ping"], buckets=5),
             heartbeat_interval=3600,
             initial_heartbeat_interval=10,
+            enable_relay_conn_data=True,
+            enable_nat_type_collection=True,
         ),
         dns=Dns(
             exit_dns=ExitDns(auto_switch_dns_ips=True),
@@ -61,7 +65,7 @@ def test_telio_features():
         ),
     )
     expected_full = TelioFeatures.from_json(
-        """{"is_test_env": false, "exit_dns": {"auto_switch_dns_ips": true}, "direct": {"providers": ["stun", "local"]}, "lana": {"prod": false, "event_path": "/"}, "nurse": {"fingerprint": "alpha", "qos": {"rtt_interval": 5, "rtt_tries": 3, "rtt_types": ["Ping"], "buckets": 5}, "heartbeat_interval": 3600, "initial_heartbeat_interval": 10}}"""
+        """{"is_test_env": false, "exit_dns": {"auto_switch_dns_ips": true}, "direct": {"providers": ["stun", "local"]}, "lana": {"prod": false, "event_path": "/"}, "nurse": {"fingerprint": "alpha", "qos": {"rtt_interval": 5, "rtt_tries": 3, "rtt_types": ["Ping"], "buckets": 5}, "heartbeat_interval": 3600, "initial_heartbeat_interval": 10, "enable_nat_type_collection": true, "enable_relay_conn_data": true}}"""
     )
     assert full_features == expected_full
     assert full_features.to_json() == expected_full.to_json()