+++ changelog.md
@@ -37,6 +37,7 @@
 * LLT-5078: Fix telio-proxy panic
 * LLT-5039: Make connection reporting unaffected by system time changes
 * LLT-4168: Fix telio-dns assert panics in debug mode
+* LLT-4856: Enable aggregator timed events
 
 <br>